requirements.txt
requests~=2.24.0
bs4~=0.0.1
selenium~=3.141.0
pandas~=1.0.5
lxml
html5lib
python-dateutil~=2.8.1
beautifulsoup4~=4.9.1
matplotlib